Refactored SQL scripts and re-factored the UHC security logic to separate file
